\name{hgu2beta7}
\alias{hgu2beta7}
\title{Bioconductor annotation data package}
\description{
  
The annotation package was built using a downloadable R package -
AnnBuilder (download and build your own) from www.bioconductor.org using
the following public data sources: 

  
	LocusLink:\url{ftp://ftp.ncbi.nih.gov/refseq/LocusLink/LL_tmpl.gz}. Built: February 16, 2005

	Gene Ontology:\url{http://www.godatabase.org/dev/database/archive/daily/go_daily-termdb.rdf-xml.gz}. Built: http://www.godatabase.org/dev/database/archive/daily/go_daily-termdb.rdf-xml.gz

	Golden Path:\url{http://hgdownload.cse.ucsc.edu/goldenPath/hg17/database/}. Built: hg17

	KEGG:\url{ftp://ftp.genome.ad.jp/pub/kegg/pathways}. Built: Release 33.0 (January 2005)

	Wed Feb 16 23:32:20 2005

	hgu2beta7

The function hgu2beta7() provides information about the binary data files}
\keyword{datasets}

