\name{hgu2beta7ORGANISM}
\alias{hgu2beta7ORGANISM}
\title{A character string for the name of the organism}
\description{

  This is a character string for the name of the organism of
  concern. The name of the organism can be viewed by typing the name of
  the binary file that will be loaded when the package is loaded by
  calling library(hgu2beta7)
}
\details{
  Although the name of the package is offten suggestive of the organism
  the package concerns, having an environment object for organism name
  make it easier to get the name of the organism programmatically.
  
  Package built Wed Feb 16 23:32:20 2005
}
\examples{
  hgu2beta7ORGANISM
}
\keyword{datasets}

